\name{urban.context-package}
\alias{urban.context}
\docType{package}
\title{
Urban context
}
\description{
For a selected point in x-y coordinates, subset and summarise points found in an specified radius. 
}
\details{
\tabular{ll}{
Package: \tab urban.context\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-10-24\cr
License: \tab GPL-2\cr
}
}
\author{
Ricardo Ochoa Sosa

Maintainer: Ricardo Ochoa Sosa <ochoasosaricardo@gmail.com>
}
\references{
None.
}

\keyword{urban}
\keyword{spatial}
